Fix angle between robots

+++ roboteam_ai/src/skills/DefendOnRobot.h
@@ -20,7 +20,7 @@ class DefendOnRobot : public Skill {
         using goType = control::ControlGoToPos::GoToType;
 
         control::ControlGoToPos goToPos;
-        float angleBetweenRobots;
+        double angleBetweenRobots;
 
         int amountOfCycles{};
     protected:
